Colourize icons in object explorer to match tab icons
